requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0
datasets==2.19.0
rank_bm25==0.2.2
transformers==4.41.2
accelerate==0.30.1
numpy<2.0.0
pandas
scikit-learn
torch